src/agentic_platform/agent/strands_glue_athena/requirements.txt
strands-agents>=1.3.0
fastapi>=0.115.11
uvicorn>=0.34.0
boto3>=1.37.13
pyjwt>=2.10.1
cryptography>=44.0.2
pydantic>=2.10.6
python-Levenshtein>=0.25.0
thefuzz>=0.22.1
